Introduce `css_bundle` variable for new `hexo-theme-meteor` feature.
The `css_bundle` is a new variable introduced by
https://github.com/meteor/hexo-theme-meteor/pull/56 and allows the selection
of a specific bundle, with the correct CSS color overrides, within the theme.

This avoids the need for non-Meteor deployments to use a deployment specific
`assets/theme-colors.less` file, though that file should still remain
to enable the ability to do deployment-specific overrides in the future.